requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.2
elevenlabs==2.14.0
openai==1.3.7
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2
nltk==3.8.1
textblob==0.17.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
celery==5.3.4
redis==5.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
